# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Proxy Tejelanas Vivi",
        "description": "API proxy para consultar información de Tejelanas Vivi con CORS habilitado",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor proxy local"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener lista de productos",
                "operationId": "072fd15d3be55577bb08c19d19f1e75c",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Productos obtenidos exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Product"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "eda0d526a50c0273ae224eb766440226",
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener un producto específico",
                "operationId": "e607d06420046014759462af284aeb59",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto obtenido exitosamente"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar un producto",
                "operationId": "d3addbe72313169741e652404694ef9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar un producto",
                "operationId": "bbf520f1a23d6a5f3dfb12813b5930d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Producto de Tejelanas Vivi",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lana Natural Merino"
                    },
                    "description": {
                        "type": "string",
                        "example": "Lana natural de alta calidad"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 15990.5
                    },
                    "stock": {
                        "type": "integer",
                        "example": 25
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "weight": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "color": {
                        "type": "string",
                        "example": "Azul marino"
                    },
                    "material": {
                        "type": "string",
                        "example": "100% Lana Merino"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token de autenticación Bearer. Tokens válidos para pruebas: tejelanas_admin_token_2025, development_token_123, test_token_456"
            }
        }
    },
...
}